SSA: - Support decimal \pos - Support ".5" in decimal numbers - Read the starting position of \move as \pos
Improves [gg]_Goku_Sayonara_Zetsubou_Sensei_-_02_[D284BF24].mkv

+++ Subtitles/SubParsing.h
@@ -36,7 +36,7 @@ extern "C"
 	unsigned marginL, marginR, marginV;
 	NSMutableArray *spans;
 	
-	int posX, posY;
+	float posX, posY;
 	UInt8 alignH, alignV, wrapStyle, render_complexity;
 	BOOL positioned;
